extract minimal json parser into srp microcrate
