+++ crates/bitnet-gpu-hal/Cargo.toml
@@ -13,6 +13,7 @@ rust-version.workspace = true
 
 [dependencies]
 bitnet-tokenizer-text-core = { path = "../bitnet-tokenizer-text-core", version = "0.2.1-dev" }
+bitnet-minimal-json-core = { path = "../bitnet-minimal-json-core", version = "0.2.1-dev" }
 bitnet-http-auth-core = { path = "../bitnet-http-auth-core", version = "0.2.1-dev" }
 bitnet-layer-index-core = { path = "../bitnet-layer-index-core", version = "0.2.1-dev" }
 tokio = { workspace = true }
