+++ crates/bitnet-minimal-json-core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "bitnet-minimal-json-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP microcrate for minimal flat JSON field extraction"
+rust-version.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
